requirements.txt
flask
python-dotenv
requests
google-generativeai
Pillow
python-docx
openpyxl
python-pptx
PyPDF2
